pyproject.toml
[project]
name = "medireply-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "jinja2>=3.1.6",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
